app/requirements.txt
pydantic_settings
fastapi[standard]
httpx
asyncpg
valkey

fastmcp

langchain-core
langchain-ollama
langchain-text-splitters
langchain-mcp-adapters

langgraph
langgraph-checkpoint-redis

arxiv
colorlog
fuzzyfinder
python-dateutil
pymupdf
pymupdf4llm